Setting PC Speaker frequency now uses integer division

+++ src/kernel/drivers/x86/pc/pcspeaker.c
@@ -14,7 +14,7 @@ void pcspeaker_set_freq (UINT freq)
 {
     FUNC_ENTRY ("frequency Hz: %x", freq);
 
-    U16 value = (U16)((double)PIT_BASE_CLOCK_FREQ_HZ / freq);
+    U16 value = (U16)(PIT_BASE_CLOCK_FREQ_HZ / freq);
     pit_set_speaker_counter (value);
 }
